requirements.txt
numpy==1.22.2
pandas==1.2.4
plotly==5.3.1
python-dotenv==0.20.0
scikit_learn==1.1.1
streamlit==1.9.2
